Extract waveform state selection helpers
